Fix LJ_MAX_JSLOTS assertion in rec_check_slots().
Thanks to Yichun Zhang.

(cherry-picked from commit 1c89933f129dde76944336c6bfd05297b8d67730)

This patch is the predecessor for the commit
944d32afd6ddd9dbac1cddf64bf81333efeb9e30 ("Add missing LJ_MAX_JSLOTS
check.") It tries to fix the issue, when `J->baseslot == LJ_MAX_JSLOTS`,
that leads to the assertion failure. Since the predecessor patch,
there are no places, that can lead to the condition failure, since we
always check that new baseslot + framesize (+ vargframe) >=
`LJ_MAX_JSLOTS`. As far as the minimum framesize is 1 (see
<src/lj_parse.c> for details), we can't obtain this assertion failure.
This patch is added for consistency with the upstream.

Since the predecessor patch fixes the issue, there is no new test case
to add.

Sergey Kaplun:
* added the description for the problem

Part of tarantool/tarantool#8825

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>