Fix IR_BUFPUT assembly.
Thanks to Peter Cawley.

(cherry picked from commit 58d0dde0a2df49abc991decbabff15230010829a)

When recording IR_BUFPTR the special variable holds -1 value to mark
that argument to be stored is not a single character. Otherwise, it can
be stored directly in a register, and this character is used to reset
the hint via the aforementioned variable at first. For the systems with
signed `char` values, the case with the character being equal to \255
produces a false positive check and leads to either invalid value
storing or even LuaJIT crash, since the variable with hint still holds
-1 value.

This patch changes the flag value to -129 to avoid collisions with
any one byte values.

Sergey Kaplun:
* added the description and the test for the problem

Reviewed-by: Igor Munkin <imun@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>