FFI: Add tonumber() specialization for failed conversions.
Contributed by Javier Guerra Giraldez.

(cherry picked from commit 02b521981a1ab919ff2cd4d9bcaee80baf77dce2)

When `tonumber()` is recorded (as a part of a trace) for a cdata
argument that can't be converted to a number, the `nil` value is recorded
as the yielded result. But without special check on trace for cdata type
this `nil` will be returned for another type of cdata that can be
converted.

This patch adds the corresponding check for recording of failed cdata
conversions.

Sergey Kaplun:
* added the description and the test for the problem

Resolves tarantool/tarantool#7655
Part of tarantool/tarantool#7230

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>