Fix os.date() for wider libc strftime() compatibility.
Thanks to Jesper Lundgren.

(cherry picked from commit fc63c938b522e147ea728b75f385728bf4a8fc35)

When `strftime()` returns empty result (for example, for "%p" on some
locales or when LuaJIT is built with musl's `strftime()` and format
string is invalid), the `os.date()` endlessly retries to format the
result and reallocates the buffer for the resulting string. This leads
to OOM.

This patch limits the amount of retries by 4.

Sergey Kaplun:
* added the description and the test for the problem

Part of tarantool/tarantool#8069

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>