Prevent integer overflow while parsing long strings.
(cherry-picked from commit 16e5605eec2e3882d709c6b123a644f6a8023945)

This commit fixes a possible integer overflow of the separator's length
counter during parsing long strings. It may lead to the fact, that the
parser considers a string with unbalanced long brackets to be correct.
Since it is pointless to parse too long string separators in the hope,
that the string is correct, just use the hardcoded limit (2 ^ 25 is
enough).

Be aware that this limit is different for Lua 5.1.

We can't check the string overflow itself without a really large file,
because the ERR_MEM error will be raised, due to the string buffer
reallocations during parsing. Keeping such a huge file in the repo is
pointless, so just check that we don't parse long strings after the
aforementioned separator length.

Sergey Kaplun:
* added the description and the test for the problem

Part of tarantool/tarantool#8825

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>