Add stricter check for print() vs. tostring() shortcut.
(cherry picked from 46a1b268eb0534182eda0447303c344a071632fe)

`print()` function uses `tostring(`) to convert all incoming values to
strings. So, if an element to convert has a `__tostring` metamethod it
must be used. In LuaJIT the `print()` base library function has a
shortcut to detect fast path for strings and numbers. This shortcut
doesn't check a metatable presence for numbers, so the `__tostring`
metamethod is ignored.

This patch adds the corresponding check for the metatable presence on
the base number type, when the aforementioned shortcut is set.

Sergey Kaplun:
* added the description and the test for the problem

Part of tarantool/tarantool#6548

Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>