ARM64: Fix {AHUV}LOAD specialized to nil/false/true.
Reported by caohongqing.

(cherry picked from commit 5bf0da3d7c02f9959fa3a9fb721e0565137b70c8)

If there is high register pressure, and there are almost all registers
in use during the aforementioned assembling, the same register is chosen
as the one holding the given stack slot and the one holding the constant
value for the type comparison. As the result we get the following
assertion guard check in assembly:
| cmp   x0, x0, lsr #32
| bne   ->0
Which is always false.

This happens because the `tmp` register (used for loading constant type
to compare in assertion guard) is scratched from `gpr` register set, but
not from `allow` set, which is used during picking the register for slot
loading (at the beginning `allow` and `gpr` sets are the same).

This patch changes `allow` set to `gpr` to fix the issue.

Sergey Kaplun:
* added the description and the test for the problem

Part of tarantool/tarantool#8069

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>