Remove pow() splitting and cleanup backends.
(cherry-picked from commit b2307c8ad817e350d65cc909a579ca2f77439682)

The JIT engine tries to split b^c to exp2(c * log2(b)) with an attempt
to rejoin them later for some backends. It adds a dependency on C99
exp2() and log2(), which aren't part of some libm implementations.
Also, for some cases for IEEE754 we can see, that exp2(log2(x)) != x,
due to mathematical functions accuracy and double precision
restrictions. So, the values on the JIT slots and Lua stack are
inconsistent.

This patch removes the splitting of pow operator, so IR_POW is emitting
for all cases (except for the power of 0.5, which is replaced with the
sqrt operation).

Also this patch does some refactoring:

* Functions `asm_pow()`, `asm_mod()`, `asm_ldexp()`, `asm_div()`
  (replaced with `asm_fpdiv()` for CPU architectures) are moved to
  <src/lj_asm.c> as far as their implementation is generic for all
  architectures.
* Fusing of IR_HREF + IR_EQ/IR_NE was moved to `asm_fuseequal()`.
* `lj_vm_exp2()` subroutine and `IRFPM_EXP2` are removed as no longer
  used.
* Also, it adds an assertion that `IR_CNEW`, `IR_CNEWI` are not emitted
  without FFI.

Sergey Kaplun:
* added the description and the test for the problem

Part of tarantool/tarantool#8825

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>