Improve assertions.
(cherry-picked from commit 8ae5170cdc9c307bd81019b3e014391c9fd00581)

This commit refactors assertions used in the LuaJIT. It introduces the
new module <src/lj_assert.c> with the `lj_assert_fail()` implementation.
Wrappers of this function are used across the whole code base. Each
macro wrapper is defined in the corresponding module and gets global
state (if possible) from its environment to be passed inside the
assertion. For now, the global state is unused, but later it may be
used for dumping the VM state.

Sergey Kaplun:
* added the description for the feature

Part of tarantool/tarantool#8825

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>