Fix handling of errors during snapshot restore.
(cherry picked from commit 12ab596997b9cb27846a5b254d11230c3f9c50c8)

When an error is raised during a snapshot restore, the `err_unwind()`
skips the correct cframe that should stop unwinding. The reason is this
frame is the pseudo C frame for error handling without an actual frame
and the special negative value of `cframe_nres()` for this frame isn't
set.

This patch sets `cframe_nres()` for a cframe that contains a snap
restoration to `-2*LUAI_MAXSTACK` to guarantee that an error will be
always caught here.

Sergey Kaplun:
* added the description and the test for the problem

Part of tarantool/tarantool#7230

Reviewed-by: Sergey Ostanevich <sergos@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>