Fix lua_yield() from C hook.
Reported by Jason Carr.

(cherry picked from commit dd5032ed844c56964347c7916db66b0eb11d8091)

When we call `lua_yield()` from the C hook, an additional continuation
frame is added. This frame contains a continuation function, PC where we
should return, a thread GC object to continue, and the frame type and
size (see details in <src/lj_frame.h>). For non-GC64 mode, when we set
the GC thread on the Lua stack, the stack top isn't incremented, so the GC
thread overwrites the PC to return. For the GC64 mode the increment is
missing before setting frame type and size.

This patch fixes the behaviour by adding the missing slot
incrementation. Also, it hardens the conditions of using
`lj_err_throw()`, according to the availability of external unwinder.

The behaviour for the GC64 mode is still wrong due to a miscalculation
of the slot of the GC thread object. This will be fixed in the next
commit.

Sergey Kaplun:
* added the description and the test for the problem

Part of tarantool/tarantool#8516

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>