test: force enable assert checks in release build
This patch fixes the regression introduced in scope of
5a61e1ab54b5c66bfebd836db1ac47996611e065 ('misc: add C and Lua API for
platform metrics'). As a result of the patch release build
was failed according to -Werror compiler flag and unused variables that
used only for assertions checks.

Force #undef NDEBUG directive leaves asserts on and allows not disable
newly added tests for release build.

Follows up tarantool/tarantool#5187