build: preserve the original build system
Since the build machinery is going to be ported to CMake there would be
Makefile names clashing. This change renames the original build system
(and a couple of auxiliary files requiring configuring) to keep all this
machinery working.

As a result of these changes one need to explicitly specify the Makefile
in the build command:
| make -f Makefile.original <options>

Needed for tarantool/tarantool#4862

Reviewed-by: Timur Safin <tsafin@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>