test: fix warnings found with luacheck in misclib*
Since the regular static analysis has not been enabled for the test
chunks in LuaJIT repository yet, the tests for recently implemented
features still produce luacheck warnings.

The most of the issues are fixed in scope of the commit
8fc103fb1a21c28185a1942e75d8d9485e3aade7 ('test: fix warnings spotted by
luacheck') and this patch fixes the remaining ones.

Resolves tarantool/tarantool#5631
Part of tarantool/tarantool#4862
Part of tarantool/tarantool#5470
Follows up tarantool/tarantool#5187

Reported-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Timur Safin <tsafin@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>